web code refactoring
