+++ src/component/DocumentationNav.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+import { NavLink } from "react-router-dom";
+
+const DocumentationNav = () => {
+  return (
+    <>
+      <NavLink to="/html" className="">
+        HTML
+      </NavLink>
+      <NavLink to="/css" className="">
+        CSS
+      </NavLink>
+      <NavLink to="/js" className="">
+        JS
+      </NavLink>
+      <NavLink to="/react" className="">
+        React
+      </NavLink>
+    </>
+  );
+};
+
+export default DocumentationNav;
