+++ src/index.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 
-import "./index.scss";
-import App from "./App";
+import "./scss/index.scss";
+import {App} from "./App";
 
 ReactDOM.render(
   <React.StrictMode>
