--- src/views/AboutMe.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const AboutMe = () => {
-  return (
-    <div>AboutMe</div>
-  )
-}
-
-export default AboutMe
