+++ src/views/FirstMenuSubpages/Documentation.jsx
@@ -1,9 +1,20 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
 
+import HomeButton from "../../component/HomeButton";
+import Svg from "../../component/Svg";
+import ThemeButton from "../../component/ThemeButton";
+
 const Documentation = () => {
   return (
     <div>
+      <header>
+        <nav className="hornaNav navigation">
+          <Svg />
+          <HomeButton />
+          <ThemeButton />
+        </nav>
+      </header>
       <div className="appSkuska">
         <div className="containerSkuska">
           <NavLink to="/html" className="item">
