+++ src/views/FirstMenuSubpages/Portfolio.jsx
@@ -1,0 +1,21 @@
+import React from "react";
+import HomeButton from "../../component/HomeButton";
+import Svg from "../../component/Svg";
+import ThemeButton from "../../component/ThemeButton";
+
+const Portfolio = () => {
+  return (
+    <>
+      <header>
+        <nav className="hornaNav">
+          <Svg />
+          <HomeButton />
+          <ThemeButton />
+        </nav>
+      </header>
+      Portfolio !
+    </>
+  );
+};
+
+export default Portfolio;
