--- src/views/Other.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Other = () => {
-  return <div>Other</div>;
-};
-
-export default Other;
