--- src/views/Portfolio.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-// import { Routes, Route, NavLink } from "react-router-dom";
-
-const Portfolio = () => {
-  return <div>Portfolio</div>;
-};
-
-export default Portfolio;
